Update CheckPoint
Created a dictionary to have an index position for each column name.  Pulled in all columns that would possibly be needed, they can be reviewed again and more removed as we go through it.  I wanted to pull in the initial information and not leave out anything we might need/want.